+++ FbonizziMonoGame/FbonizziMonoGame/FbonizziMonoGame.csproj
@@ -41,16 +41,17 @@
     <Compile Include="Achievements\PopupAchievement.cs" />
     <Compile Include="AfterGameManagement\AfterGameManager.cs" />
     <Compile Include="AfterGameManagement\ProtipsShower.cs" />
+    <Compile Include="Drawing\IScreenTransformationMatrixProvider.cs" />
     <Compile Include="Sprites\AnimationsManager.cs" />
     <Compile Include="Assets\SplashScreenLoader.cs" />
     <Compile Include="Extensions\CollectionsExtensions.cs" />
-    <Compile Include="ILocalizedStringsRepository.cs" />
-    <Compile Include="ITextFileLoader.cs" />
+    <Compile Include="PlatformAbstractions\ILocalizedStringsRepository.cs" />
+    <Compile Include="PlatformAbstractions\ITextFileLoader.cs" />
     <Compile Include="Assets\TexturePackerCustomImporter.cs" />
     <Compile Include="Assets\SpriteDescription.cs" />
     <Compile Include="Drawing\BoxingViewportAdapter.cs" />
     <Compile Include="Drawing\Camera2D.cs" />
-    <Compile Include="Drawing\ScalingViewportAdapter.cs" />
+    <Compile Include="Drawing\ScalingMatrixProvider.cs" />
     <Compile Include="Exceptions\AchievementCheckFuncNotDefinedException.cs" />
     <Compile Include="Extensions\ColorExtensions.cs" />
     <Compile Include="Extensions\RectangleExtensions.cs" />
@@ -60,7 +61,7 @@
     <Compile Include="Drawing\DrawingInfos.cs" />
     <Compile Include="Extensions\Texture2DExtensions.cs" />
     <Compile Include="Extensions\TimeSpanExtensions.cs" />
-    <Compile Include="IFbonizziGame.cs" />
+    <Compile Include="PlatformAbstractions\IFbonizziGame.cs" />
     <Compile Include="Input\GamePadEventArgs.cs" />
     <Compile Include="Input\GamePadListener.cs" />
     <Compile Include="Input\IInputListener.cs" />
@@ -76,7 +77,7 @@
     <Compile Include="UI\RateMeDialog.cs" />
     <Compile Include="UI\ButtonWithText.cs" />
     <Compile Include="UI\ColorChangingObject.cs" />
-    <Compile Include="IWebPageOpener.cs" />
+    <Compile Include="PlatformAbstractions\IWebPageOpener.cs" />
     <Compile Include="UI\InMemoryLocalizedStringsRepository.cs" />
     <Compile Include="UI\RateMeDialogStrings.cs" />
     <Compile Include="UI\ScalingObject.cs" />
@@ -85,7 +86,7 @@
     <Compile Include="Particles\Particle.cs" />
     <Compile Include="Particles\ParticleGenerator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ISettingsRepository.cs" />
+    <Compile Include="PlatformAbstractions\ISettingsRepository.cs" />
     <Compile Include="Sprites\Sprite.cs" />
     <Compile Include="Sprites\SpriteAnimation.cs" />
     <Compile Include="Extensions\StringBuilderExtensions.cs" />
